EncontristaForm and Main component created and routes changed
