+++ src/routes.js
@@ -7,6 +7,7 @@ import LoginPage from "./pages/Login";
 import { AuthProvider } from "./services/auth";
 import { PrivateRoute } from "./PrivateRoute";
 import { EncontreirosFrom } from "./pages/EncontreirosFrom";
+import { EncontristasForm } from "./pages/EncontristasForm";
 
 export default function Routes() {
   return (
@@ -17,11 +18,7 @@ export default function Routes() {
           <PrivateRoute path="/list" exact component={List} />
           <Route path="/autentication" exact component={LoginPage} />
           <Route path="/encontreiro" exact component={EncontreirosFrom} />
-          <Route
-            path="/encontrista"
-            exact
-            component={() => <h1>Em construçao! ☺ </h1>}
-          />
+          <Route path="/encontrista" exact component={EncontristasForm} />
           <Route path="*" component={() => <h1>Page not found</h1>} />
         </Switch>
       </BrowserRouter>
